feat : Implemented Controller, Repository and Service for Rental. fix : Corrected DTO with BigDecimal

+++ back-end/api/src/main/java/com/chatop/api/dto/RentalDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -14,8 +15,8 @@ import java.time.LocalDateTime;
 public class RentalDto {
     private Integer id;
     private String name;
-    private Double surface;
-    private Double price;
+    private BigDecimal surface;
+    private BigDecimal price;
     private String picture;
     private String description;
     private Integer owner_id; // On renvoie juste l'ID, pas tout l'objet User
